% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.R
\name{calfsbm_cv}
\alias{calfsbm_cv}
\title{Cross-validation function for CALF-SBM variable selection}
\usage{
calfsbm_cv(network, K, folds = 4, cluster_intercept = FALSE)
}
\arguments{
\item{network}{List object containing adjacency (A), similarity (dis), 
and covariates (X)}

\item{K}{Number of clusters}

\item{folds}{Number of folds (default = 4)}

\item{cluster_intercept}{Logical (default = \code{FALSE}) indicating 
whether the intercept term should be a matrix or a single value}
}
\value{
BIC of cross-validated model
}
\description{
Cross-validation function for CALF-SBM variable selection
}
