% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.R
\name{calfsbm_em_exact}
\alias{calfsbm_em_exact}
\title{Function to perform an exact EM algorithm on network data}
\usage{
calfsbm_em_exact(
  network,
  K,
  cluster_intercept = FALSE,
  verbose = TRUE,
  find_cov = FALSE
)
}
\arguments{
\item{network}{List object containing similarity matrix dis,}

\item{K}{Number of clusters}

\item{cluster_intercept}{Logical (default = \code{FALSE}) indicating 
whether the intercept term should be a matrix or a single value}

\item{verbose}{Logical indicating the verbosity of the function}

\item{find_cov}{Logical (default = \code{FALSE}) indicating whether the covariance
matrix accounting for the uncertainty in the z should be found, will return 
observed-data covariance otherwise. The complete variance estimation 
can be computationally intensive if chosen}
}
\value{
Soft clustering estimate z, beta0 and beta estimates, 
log-likelihood, GLM output, and computing time
}
\description{
Function to perform an exact EM algorithm on network data
}
