% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.R
\name{calfsbm_q}
\alias{calfsbm_q}
\title{Fixed point EM iteration to be used for squarem}
\usage{
calfsbm_q(beta, network, z, cluster_intercept = FALSE)
}
\arguments{
\item{beta}{description}

\item{network}{List object containing the following attributes:
dis: Similarity matrix
A: Adjacency matrix}

\item{z}{Clustering vector}

\item{cluster_intercept}{Logical indicating whether the intercept should be 
a single value or a matrix corresponding to cluster label}
}
\value{
Vector of updated betas
}
\description{
Fixed point EM iteration to be used for squarem
}
