% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.R
\name{calfsbm_q_exact}
\alias{calfsbm_q_exact}
\title{Exact fixed-point EM iteration to be used for squarem}
\usage{
calfsbm_q_exact(beta, network, z, cluster_intercept = FALSE)
}
\arguments{
\item{beta}{Vector containing initial beta parameter estimates}

\item{network}{List object containing the following attributes:
dis: Similarity matrix
A: Adjacency matrix}

\item{z}{Soft clustering matrix, with each row corresponding to a node,
and each column corresponding to a cluster}

\item{cluster_intercept}{Logical indicating whether the intercept should be 
a single value or a matrix corresponding to cluster label}
}
\value{
Vector of updated betas
}
\description{
Exact fixed-point EM iteration to be used for squarem
}
