% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.R
\name{cammsbm_cv}
\alias{cammsbm_cv}
\title{Cross-validation function for CAMM-SBM variable selection}
\usage{
cammsbm_cv(network, K, folds = 5)
}
\arguments{
\item{network}{List object containing adjacency (A), similarity (dis), 
and covariates (X)}

\item{K}{Number of clusters}

\item{folds}{Number of folds (default = 5)}
}
\value{
Predictive pseudo-likelihood of cross-validated model
}
\description{
Cross-validation function for CAMM-SBM variable selection
}
