% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.R
\name{cammsbm_nimble}
\alias{cammsbm_nimble}
\title{Function to run MCMC version of CAMM-SBM model}
\usage{
cammsbm_nimble(
  network,
  K,
  nsim = 2000,
  burnin = 1000,
  thin = 2,
  nchain = 2,
  beta_scale = 10
)
}
\arguments{
\item{network}{List object containing adjacency (A), similarity (dis), 
and covariates (X)}

\item{K}{Number of clusters}

\item{nsim}{Number of simulations}

\item{burnin}{Number of burn-in MCMC iterations to discard}

\item{thin}{Thinning rate of post-burn-in iterations}

\item{nchain}{Number of chains to run}

\item{beta_scale}{Standard deviation for the beta parameters}
}
\value{
WAIC of the fitted model
}
\description{
Function to run MCMC version of CAMM-SBM model
}
