% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.R
\name{cammsbm_plem}
\alias{cammsbm_plem}
\title{Pseudo-likelihood EM algorithm implementation}
\usage{
cammsbm_plem(network, K, max_iter = 20, tol = 1e-04, verbose = TRUE)
}
\arguments{
\item{network}{List object containing adjacency (A), similarity (dis), 
and covariates (X)}

\item{K}{Number of clusters}

\item{max_iter}{Numeric (default = 20) indicating 
the maximum number of iterations to run (in the case of clustering 
configuration alternating between iterations)}

\item{tol}{Positive numeric indicating the gain in log-likelihood for which
the algorithm will terminate if the improvement is less than}

\item{verbose}{Verbosity (default = \code{TRUE})}
}
\value{
List of estimated node membership, betas, and log-pseudo-likelihood
}
\description{
Pseudo-likelihood EM algorithm implementation
}
