% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.R
\name{cammsbm_q_vem}
\alias{cammsbm_q_vem}
\title{Fixed-point variational EM iteration to be used for squarem}
\usage{
cammsbm_q_vem(beta_gamma_vec, network, K)
}
\arguments{
\item{beta_gamma_vec}{Vector containing initial beta parameter estimates
and gamma estimates, flattened into a single vector}

\item{network}{List object containing adjacency (A), similarity (dis), 
and covariates (X)}

\item{K}{Number of clusters}
}
\value{
Vector of updated betas and gammas
}
\description{
Fixed-point variational EM iteration to be used for squarem
}
