% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{convergence_procedure}
\alias{convergence_procedure}
\title{(Obsolete) Check MCMC for convergence}
\usage{
convergence_procedure(z, group, niter, i, K)
}
\arguments{
\item{z}{Node membership from last five iterations}

\item{group}{}

\item{niter}{Total iterations initially planned}

\item{iter}{The current iteration of the Gibbs sampler}
}
\value{
Simulated coefficients
}
\description{
Checks MCMC by determining if node membershpi is stable
}
