% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{find_k_best_bic}
\alias{find_k_best_bic}
\title{(Obsolete) K-selection Function}
\usage{
find_k_best_bic(p, alpha, beta0, beta, niter, A, S_ij)
}
\arguments{
\item{p}{Number of clusters to test (from 1 to p)}

\item{alpha}{Vector of prior probabilities of being in each group}

\item{beta0}{Prior intercept for logistic regression}

\item{beta}{Prior parameters for logistic regression, in matrix form}

\item{niter}{Number of iterations to run}

\item{A}{Adjacency matrix}

\item{S_ij}{Distance matrix}
}
\value{
List containing the estimated beta, node membership, and history
}
\description{
Function that uses BIC to find optimal number of clusters
}
\note{
Deprecated
}
