% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{find_K_optimal}
\alias{find_K_optimal}
\title{Latentnet Cluster Finder}
\usage{
find_K_optimal(p, AA, sample_size, burnin, by)
}
\arguments{
\item{p}{Number of clusters. The function will iterate from 1 to p}

\item{AA}{Network object containing adjacency and covariate information}

\item{sample_size}{Number of non-rejected MCMC samples}

\item{burnin}{Number of MCMC samples to allow the algorithm to converge}

\item{by}{Number of samples to skip after burn-in}
}
\value{
A dataframe with the AIC and BIC values for each K value
}
\description{
Function to find the optimal number of groups using latentnet package
}
