% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.R
\name{forward_stepwise_vem}
\alias{forward_stepwise_vem}
\title{Function to perform forward stepwise variable selection}
\usage{
forward_stepwise_vem(network, K, cluster_intercept = FALSE, cv = TRUE)
}
\arguments{
\item{network}{List object containing adjacency (A), similarity (dis), 
and covariates (X)}

\item{K}{Number of clusters}

\item{cluster_intercept}{Whether to use a matrix of beta_0 values 
(default = FALSE)}

\item{cv}{Whether to use cross-validation (default = TRUE)}
}
\value{
Vector of covariate indices which correspond to the optimal model
}
\description{
Function to perform forward stepwise variable selection
}
