% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{gelman}
\alias{gelman}
\title{Gelman diagnostic}
\usage{
gelman(results, as_matrix = TRUE, n_chains = 3)
}
\arguments{
\item{results}{A matrix, where each column represents a chain}

\item{as_matrix}{Boolean indicating whether matrix or array input is used}

\item{n_chains}{Number of MCMC chains}
}
\value{
Gelman-Rubin diagnostic (ideal value = 1)
}
\description{
Takes matrix and converts it to a useful format
}
\note{
Function has been superseded by the return_gelman option in 
\code{calf_sbm_nimble} function
}
