% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{gen_az}
\alias{gen_az}
\title{CALF-SBM Network Generation Function}
\usage{
gen_az(n_nodes, K, m, prob, beta0, beta, sigma, spat, directed = FALSE)
}
\arguments{
\item{n_nodes}{Number of nodes in the network}

\item{K}{True number of clusters}

\item{m}{The number of covariates in X}

\item{prob}{Vector of weights of being in each group. This should be
a numeric vector of length K, and can be weights or probabilities}

\item{beta0}{Intercept term, lower values lead to higher sparsity}

\item{beta}{Matrix of within and between-cluster coefficients. Diagonal
values correspond to within-cluster effects}

\item{sigma}{Standard deviation of offset term, set to 0 if not using offset}

\item{spat}{Signal-to-noise ratio of the covariates. 0 indicates no signal}

\item{directed}{Boolean indicating whether the network should be directed}
}
\value{
list of adjacency, membership, link probability, and distance
}
\description{
Generates network using probabilities, betas, and signal strength
}
