% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{gen_az_ln}
\alias{gen_az_ln}
\title{Latentnet Network Generator}
\usage{
gen_az_ln(n_nodes, K, x_dim, betavec, noise = 1, directed = FALSE)
}
\arguments{
\item{n_nodes}{Number of nodes}

\item{K}{Number of clusters}

\item{x_dim}{Number of covariates}

\item{betavec}{Vector of intercept and slope terms}

\item{noise}{noise/signal ratio (default=1)}
}
\value{
Network as list object with characteristics
}
\description{
Function to generate network according to Handcock's RSS latentnet paper
}
