% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.R
\name{gen_factor_soft}
\alias{gen_factor_soft}
\title{Function to set up Q function in exact EM form}
\usage{
gen_factor_soft(initial_z, A, S_ij, directed = FALSE, offset = FALSE)
}
\arguments{
\item{initial_z}{n-by-K matrix of node-level clustering probabilities}

\item{A}{n-by-n adjacency matrix}

\item{S_ij}{n-by-n similarity matrix}

\item{directed}{Logical indicating whether the network should be directed}

\item{offset}{Logical indicating whether to include node-level 
random-effects terms (default = FALSE)}
}
\value{
Data frame with weights, combinations of connections and clusters
}
\description{
Function to set up Q function in exact EM form
}
\note{
Internal helper for \code{cammsbm_plem_exact} and \code{calfsbm_q_exact}
}
