% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{generate_calfsbm_network}
\alias{generate_calfsbm_network}
\title{CALF-SBM Network Generation Function}
\usage{
generate_calfsbm_network(
  n_nodes,
  K,
  n_covar,
  prob,
  beta0,
  beta,
  sigma,
  spat,
  directed = FALSE
)
}
\arguments{
\item{n_nodes}{Number of nodes in the network}

\item{K}{A positive integer indicating the true number of clusters}

\item{n_covar}{A positive integer indicating the number of node-specific
covariates to generate.}

\item{prob}{Vector of weights of being in each group. This should be
a numeric vector of length K, and can be weights or probabilities}

\item{beta0}{Intercept term, lower values correspond to higher sparsity}

\item{beta}{K-by-K matrix of within and between-cluster coefficients. 
Diagonal values correspond to within-cluster effects}

\item{sigma}{Non-negative number for the standard deviation of the random 
effects offset term, set to 0 if not using offset}

\item{spat}{Non-negative number to represent the ratio of between-cluster 
variance to within-cluster variance of the covariates. 0 indicates no signal}

\item{directed}{logical; if \code{FALSE} (default), the MCMC output is from 
an undirected network}
}
\value{
List of adjacency, membership, link probability, and distance
}
\description{
Generates network using probabilities, betas, and signal strength
}
\examples{
generate_calfsbm_network(100, 2, 2, c(0.5, 0.5), 1, 
diag(2) - 3, 0.3, 1.5)
}
