% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{generate_latentnet_network}
\alias{generate_latentnet_network}
\title{Latentnet Network Generator}
\usage{
generate_latentnet_network(
  n_nodes,
  K,
  x_dim,
  betavec,
  noise = 1,
  directed = FALSE
)
}
\arguments{
\item{n_nodes}{A positive integer representing the total number of nodes in 
the network}

\item{K}{A positive integer indicating the true number of clusters}

\item{x_dim}{A positive integer indicating the number of node-specific
covariates to generate.}

\item{betavec}{Vector of intercept and slope terms}

\item{noise}{noise/signal ratio (default=1)}

\item{directed}{logical; if \code{FALSE} (default), the MCMC output is from 
an undirected network}
}
\value{
Network as list object with characteristics
}
\description{
Function to generate network according to Handcock's RSS latentnet paper
}
\examples{
links <- generate_latentnet_network(200, 2, 2, c(1, 1))
print(find_sbm(links$A, links$z))

}
\references{
Handcock, M. S., A. E. Raftery, and T. J. M. (2007). Model-based 
clustering for social networks. Journal of the Royal Statistical Society. 
Series A 170 (2), 307–354
}
