% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{gibbs_sampler_unknown_k}
\alias{gibbs_sampler_unknown_k}
\title{(Obsolete) Unknown K Gibbs sampler}
\usage{
gibbs_sampler_unknown_k(dp, niter, A, S_ij, directed)
}
\arguments{
\item{dp}{Dirichlet process parameter (recommended 1)}

\item{niter}{Number of iterations to run}

\item{A}{Adjacency matrix}

\item{S_ij}{Distance matrix}

\item{X}{Matrix of covariates}
}
\value{
List containing the estimated beta, node membership, and history
}
\description{
Gibbs sample with CRP as initial state
}
