% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{makesymmetric}
\alias{makesymmetric}
\title{Helper Function to Make Adjacency Matrix Symmetric}
\usage{
makesymmetric(mat)
}
\arguments{
\item{mat}{The matrix to be symmetrized}
}
\value{
Symmetric matric with the same upper diagonal as the original
}
\description{
Makes a non-symmetric matrix symmetric using the lower diagonal
}
