% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{mfm_sbm}
\alias{mfm_sbm}
\title{Function for MFM-SBM Clustering}
\usage{
mfm_sbm(z, A, conc, S_ij, niter = 100)
}
\arguments{
\item{z}{node membership vector}

\item{A}{Known connectivity matrix}

\item{conc}{Concentration parameter of CRP}

\item{S_ij}{Distance}

\item{niter}{Number of iterations}
}
\value{
List of beta, z, and history of K
}
\description{
Implementation of Algorithm 1 in MFM-SBM for the model
}
\examples{
links <- gen_az(n_nodes = 50, K = 2, m = 2, prob = c(0.5, 0.5),
beta0 = 1, beta = diag(2) - 3, sigma = 0.3, spat = 0.5);
mfm_sbm(links$z, links$A, 0.8, links$dis, 1000)
}
