% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{plot_conn}
\alias{plot_conn}
\title{Visualize Link Probability vs. Actual Matrix}
\usage{
plot_conn(observed, prob, z_tru)
}
\arguments{
\item{observed}{Observed adjacency matrix}

\item{prob}{Matrix of intrinsic probabilities}

\item{z_tru}{Vector of the true node membership}
}
\value{
Graph of the connectivity and graph of true probabilities
}
\description{
Using the adjacency matrix, returns a plot of the adjacency matrix, and 
a raster of the true probabilities
}
\examples{
links <- gen_az(n_nodes = 50, K = 2, m = 2, prob = c(0.5, 0.5),
beta0 = 1, beta = diag(2) - 3, sigma = 0.3, spat = 0.5)
plot_conn(links$A, links$lp, links$z)
}
