% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{post_label_mcmc}
\alias{post_label_mcmc}
\title{Label-Switching According to Constraint}
\usage{
post_label_mcmc(mcmcSamples, K, directed = FALSE, lab = TRUE)
}
\arguments{
\item{mcmcSamples}{The set of betas}

\item{K}{Number of groups}

\item{directed}{Boolean indicating whether network is directed}

\item{lab}{Boolean indicating whether to worry about label-switching}
}
\value{
Reorganized dataframe according to apparent beta_{ii} ordering
}
\description{
Function to post-label MCMC samples by ascending order of within-cluster values
}
