% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{post_label_mcmc_samples}
\alias{post_label_mcmc_samples}
\title{Function to post-process raw MCMC output sample-by-sample}
\usage{
post_label_mcmc_samples(mcmcSamples, K, directed = FALSE)
}
\arguments{
\item{mcmcSamples}{The set of betas}

\item{K}{Number of groups}

\item{directed}{Boolean indicating whether network is directed}
}
\value{
Entire MCMC samples data reorganized according to constraint
}
\description{
Function to post-process raw MCMC output sample-by-sample
}
