% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print.cammsbm}
\alias{print.cammsbm}
\title{Print method for cammsbm objects}
\usage{
\method{print}{cammsbm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"cammsbm"} returned by 
\code{\link{cammsbm_vem}}.}

\item{...}{Additional arguments (currently none).}
}
\value{
Invisibly returns the input object.
}
\description{
Displays a concise summary of a fitted CAMM-SBM model.
}
