% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.R
\name{sem}
\alias{sem}
\title{Function to run the SEM Algorithm to estimate variance of parameters}
\usage{
sem(network, K, beta, z, obs_fisher, increment = 1e-04)
}
\arguments{
\item{network}{List object containing similarity matrix dis, 
adjacency matrix A, and matrix of covariates X}

\item{K}{Number of clusters}

\item{beta}{Fitted beta values from the EM algorithm}

\item{z}{Fitted matrix of z probabilities}

\item{obs_fisher}{Observed Fisher information from the EM fit}

\item{increment}{Small increment to calculate numerical Jacobian}
}
\value{
Covariance matrix of betas
}
\description{
Function to run the SEM Algorithm to estimate variance of parameters
}
\note{
Internal helper
}
