% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.R
\name{sem_richardson}
\alias{sem_richardson}
\title{SEM Algorithm using Richardson extrapolation}
\usage{
sem_richardson(
  network,
  K,
  beta,
  z,
  obs_fisher,
  cluster_intercept = FALSE,
  increment = 1e-04
)
}
\arguments{
\item{network}{List object containing similarity matrix dis, 
adjacency matrix A, and matrix of covariates X}

\item{K}{Number of clusters}

\item{beta}{Fitted beta values from the EM algorithm, as vector}

\item{z}{Fitted matrix of z probabilities}

\item{obs_fisher}{Observed Fisher information from the EM fit}

\item{cluster_intercept}{Logical indicating whether the intercept should be 
a single value or a matrix corresponding to cluster label}

\item{increment}{Small increment to calculate numerical Jacobian}
}
\value{
Covariance matrix of betas
}
\description{
SEM Algorithm using Richardson extrapolation
}
\note{
Internal helper
}
