% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{serialize}
\alias{serialize}
\title{Serialize Cluster Labels}
\usage{
serialize(z)
}
\arguments{
\item{z}{Vector of node membership to be renumbered from 1-K}
}
\value{
Serialized node membership
}
\description{
Number clusters from 1-K in case a cluster was deleted
}
\examples{
serialize(c(2, 2, 3, 3))
}
