% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{sim_study_K_finder}
\alias{sim_study_K_finder}
\title{(Obsolete) Function to Simulate K Frequencies}
\usage{
sim_study_K_finder(
  nsim,
  n_vals,
  K,
  p,
  x_dim,
  beta0,
  beta,
  sample_size = 2000,
  burnin = 4000,
  by = 4
)
}
\arguments{
\item{nsim}{The total number of simulations}

\item{n_vals}{Vector of the total node numbers}

\item{K}{True number of clusters, can also be vector form}

\item{p}{The maximum number of clusters to test for}

\item{x_dim}{Number of covariates in X}

\item{beta0}{Intercept term}

\item{beta}{Effect of clusters on link probability}
}
\value{
Matrix of true K vs. estimated K
}
\description{
Function to expand K optimization into a simulation study
}
