% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.R
\name{update_beta_em}
\alias{update_beta_em}
\title{Helper function to update beta according to adjacency and node membership}
\usage{
update_beta_em(
  K,
  group,
  init = NULL,
  directed = FALSE,
  cluster_intercept = FALSE
)
}
\arguments{
\item{K}{A positive integer indicating the true number of clusters}

\item{group}{Model matrix to be fitted on}

\item{init}{Starting values for logistic regression (default = \code{NULL})}

\item{directed}{logical; if \code{FALSE} (default), the MCMC output is from 
an undirected network}

\item{cluster_intercept}{Logical indicating whether the intercept term
should be a matrix or a single value}
}
\value{
beta0 and beta, with beta as a matrix
}
\description{
Helper function to update beta according to adjacency and node membership
}
\note{
Function \code{update_beta_em} is a helper in the 
in the \code{cammsbm_plem} function, deriving an estimate for beta using 
the initial clustering configuration as input
}
