% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.R
\name{update_beta_exact}
\alias{update_beta_exact}
\title{Update Beta in exact EM form}
\usage{
update_beta_exact(
  K,
  group,
  init = NULL,
  directed = FALSE,
  offset = FALSE,
  cluster_intercept = TRUE
)
}
\arguments{
\item{K}{A positive integer indicating the true number of clusters}

\item{group}{Model matrix to be fitted on}

\item{init}{Starting values for logistic regression (default = \code{NULL})}

\item{directed}{logical; if \code{FALSE} (default), the MCMC output is from 
an undirected network}

\item{offset}{Logical indicating whether to use offset term}

\item{cluster_intercept}{Logical indicating whether the intercept term
should be a matrix or a single value}
}
\value{
List object with fitted beta0 and beta, along with model diagnostics
}
\description{
Helper function to update beta according to adjacency and node membership
}
\note{
Function \code{update_beta_exact} is a helper in the 
in the \code{calfsbm_em_exact} function, deriving an estimate for beta using 
the initial clustering configuration as input
}
