% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{update_theta}
\alias{update_theta}
\title{Helper function to update theta offset in EM implementation}
\usage{
update_theta(A, S_ij, z, beta0, beta, theta)
}
\description{
Helper function to update theta offset in EM implementation
}
