% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{update_z_from_q}
\alias{update_z_from_q}
\title{Helper to update MFM-SBM by iteration}
\usage{
update_z_from_q(z, Q, A, conc, a = 1, b = 1)
}
\arguments{
\item{z}{Vector of node membership, length n}

\item{Q}{Stochastic Block Model probability matrix}

\item{A}{Network adjacency matrix, dimension n x n}

\item{conc}{Concentration parameter}

\item{a}{Shape parameter of prior beta distribution}

\item{b}{Shape parameter of prior beta distribution}
}
\value{
Updated node membership
}
\description{
Helper function for MFM-SBM Algorithm 1
}
\note{
Internal helper
}
