% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.R
\name{vec2mat}
\alias{vec2mat}
\title{Helper function to switch from a vector to a matrix representation of beta}
\usage{
vec2mat(beta, directed = FALSE)
}
\arguments{
\item{beta}{Vector of entries}

\item{directed}{Logical indicating whether network is directed}
}
\value{
Matrix of entries
}
\description{
Helper function to switch from a vector to a matrix representation of beta
}
