% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.R
\name{vem_m}
\alias{vem_m}
\title{Helper M-Step for VEM algorithm}
\usage{
vem_m(network, gamma, beta0 = NULL, beta = NULL)
}
\arguments{
\item{network}{List object containing adjacency (A), similarity (dis), 
and covariates (X)}

\item{gamma}{n x K matrix of clustering probabilities}

\item{beta0}{Optional K x K intercept matrix (default = \code{NULL})}

\item{beta}{K x K matrix (default = \code{NULL})}
}
\value{
List of fitted beta0 and beta matrices
}
\description{
Helper M-Step for VEM algorithm
}
