% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_base.R
\name{calf_sbm_nimble}
\alias{calf_sbm_nimble}
\title{Implementation of CALF-SBM}
\usage{
calf_sbm_nimble(
  links,
  nsim,
  burnin,
  thin,
  nchain,
  K,
  directed,
  offset = TRUE,
  beta_scale = 10
)
}
\arguments{
\item{links}{List of elements of the network, 
requires adjacency matrix A, matrix of covariates X, and distance matrix dis}

\item{nsim}{Total number of MCMC iterations per chain}

\item{burnin}{Number of iterations in each chain to be discarded}

\item{thin}{Post-burnin thinning parameter}

\item{nchain}{Number of MCMC chains to run}

\item{K}{Number of clusters}

\item{directed}{Boolean indicating whether to use directed network 
(default = FALSE)}

\item{offset}{Boolean indicating whether to use offset terms}

\item{beta_scale}{Prior standard deviation of beta terms}
}
\value{
List of beta, z, and history of K
}
\description{
Using MCMC parameters and number of clusters as input, return raw MCMC output
}
