% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{calf_sbm_nimble}
\alias{calf_sbm_nimble}
\title{Implementation of CALF-SBM}
\usage{
calf_sbm_nimble(
  links,
  nsim,
  burnin,
  thin,
  nchain,
  K,
  offset = TRUE,
  beta_scale = 10,
  return_gelman = FALSE
)
}
\arguments{
\item{links}{List of elements of the network, 
requires adjacency matrix A, matrix of covariates X, and distance matrix dis}

\item{nsim}{Total number of MCMC iterations per chain}

\item{burnin}{Number of iterations in each chain to be discarded}

\item{thin}{Post-burnin thinning parameter}

\item{nchain}{Number of MCMC chains to run}

\item{K}{Number of clusters}

\item{offset}{logical (default = \code{TRUE}); where \code{TRUE} 
indicates to use offset terms theta in the \code{NIMBLE} model}

\item{beta_scale}{Prior standard deviation of beta terms}

\item{return_gelman}{logical (default = \code{FALSE}); if \code{TRUE}, 
returns the Gelman-Rubin diagnostic for all beta terms}
}
\value{
List of beta, z, and history of K
}
\description{
Using MCMC parameters and number of clusters as input, return raw MCMC output
}
