% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.R
\name{calfsbm_em}
\alias{calfsbm_em}
\title{EM algorithm implementation to get variational Bayesian distribution}
\usage{
calfsbm_em(
  network,
  K,
  cluster_intercept = FALSE,
  verbose = TRUE,
  find_cov = FALSE,
  tol = 1e-06
)
}
\arguments{
\item{network}{A list object containing adjacency matrix A,}

\item{K}{Number of clusters}

\item{cluster_intercept}{Logical (default = \code{FALSE}) indicating 
whether the intercept term should be a matrix or a single value}

\item{verbose}{Verbosity (default = \code{TRUE})}

\item{find_cov}{Logical (default = \code{FALSE}) indicating whether the covariance
matrix accounting for the uncertainty in the z should be found, will return 
observed-data covariance otherwise. The complete variance estimation 
can be computationally intensive if chosen}

\item{tol}{Positive numeric indicating the gain in log-likelihood for which
the algorithm will terminate if the improvement is less than}
}
\value{
List of estimated node membership, betas, and AIC
}
\description{
EM algorithm implementation to get variational Bayesian distribution
}
