% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{calfsbm_em}
\alias{calfsbm_em}
\title{EM algorithm implementation to get variational Bayesian distribution}
\usage{
calfsbm_em(network, K, offset = TRUE, verbose = TRUE)
}
\arguments{
\item{network}{A list object containing adjacency matrix A,}

\item{K}{Number of clusters}

\item{offset}{Boolean to indicate if there should be an offset (default = TRUE)}

\item{verbose}{Verbosity (default = TRUE)}
}
\value{
List of estimated node membership, betas, and AIC
}
\description{
EM algorithm implementation to get variational Bayesian distribution
}
