% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{convergence_procedure}
\alias{convergence_procedure}
\title{(Obsolete) Check MCMC for convergence}
\usage{
convergence_procedure(z, group, niter, i, K, directed = FALSE)
}
\arguments{
\item{z}{Node membership from last five iterations}

\item{group}{Output from \code{gen_factor}}

\item{niter}{Total iterations initially planned}

\item{i}{The current iteration of the Gibbs sampler}

\item{K}{Number of clusters}

\item{directed}{logical; if \code{FALSE} (default), assumes adjacency matrix 
is undirected and symmetric}
}
\value{
Simulated coefficients
}
\description{
Checks MCMC by determining if node membership is stable
}
\note{
Deprecated
}
