% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.R
\name{find_elbo}
\alias{find_elbo}
\title{Function to find the ELBO for the CAMM-SBM model. As currently constructed,
the function will actually report the negative of the ELBO to be compatible
with the optim() function, which only minimizes objective functions}
\usage{
find_elbo(beta_vec, network, z, cluster_intercept = TRUE)
}
\arguments{
\item{beta_vec}{beta_0 an beta parameters flattened into 
a combined vector}

\item{network}{}

\item{z}{Matrix of clustering probabilities found in the E-Step}
}
\value{
Observed ELBO with the given parameters
}
\description{
Function to find the ELBO for the CAMM-SBM model. As currently constructed,
the function will actually report the negative of the ELBO to be compatible
with the optim() function, which only minimizes objective functions
}
