% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{find_K_optimal}
\alias{find_K_optimal}
\title{Latentnet Cluster Finder}
\usage{
find_K_optimal(p, AA, sample_size, burnin, by)
}
\arguments{
\item{p}{Number of clusters. The function will iterate from 1 to p}

\item{AA}{Network object (from network package) containing adjacency and 
covariate information}

\item{sample_size}{Number of non-rejected MCMC samples}

\item{burnin}{Number of initial MCMC samples to be discarded}

\item{by}{Number of samples to skip after burn-in period}
}
\value{
A dataframe with the AIC and BIC values for each K value
}
\description{
Function to find the optimal number of groups using latentnet package
}
\examples{
net <- network::network(m <- matrix(rbinom(25, 1, 0.4), 5, 5));
find_K_optimal(6, net, 10000, 2000, 5)
}
