% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{find_K_optimal}
\alias{find_K_optimal}
\title{Latentnet Cluster Finder}
\usage{
find_K_optimal(p, AA, sample_size, burnin, by)
}
\arguments{
\item{p}{Number of clusters. The function will iterate from 1 to p}

\item{AA}{Network object (from network package) containing adjacency and 
covariate information}

\item{sample_size}{Number of non-rejected MCMC samples}

\item{burnin}{Number of initial MCMC samples to be discarded}

\item{by}{Number of samples to skip after burn-in period}
}
\value{
A dataframe with the AIC and BIC values for each K value
}
\description{
Function to find the optimal number of groups using latentnet package
}
\examples{
\dontrun{
set.seed(123)
net <- network::network(m <- matrix(stats::rbinom(2500, 1, 0.4), 50, 50))
find_K_optimal(4, net, 1000, 200, 1)
}

}
\references{
Handcock, M. S., A. E. Raftery, and T. J. M. (2007). Model-based 
clustering for social networks. Journal of the Royal Statistical Society. 
Series A 170 (2), 307–354
}
