% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{find_sbm}
\alias{find_sbm}
\title{Condense Adjacency into SBM with Known Node Membership}
\usage{
find_sbm(A, z)
}
\arguments{
\item{A}{The observed adjacency matrix}

\item{z}{The true node membership (values assumed to be in [1, K])}
}
\value{
A K x K matrix with the observed density of each block
}
\description{
Function to calculate the implied stochastic block probabilities
}
\examples{
set.seed(123)
links <- gen_az(n_nodes = 50, K = 2, m = 2, prob = c(0.5, 0.5),
beta0 = 1, beta = diag(2) - 3, sigma = 0.3, spat = 0.5)
find_sbm(links$A, links$z)
}
