% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{gen_az}
\alias{gen_az}
\title{CALF-SBM Network Generation Function}
\usage{
gen_az(
  n_nodes,
  K,
  x_dim,
  prob,
  beta0,
  beta,
  sigma,
  spat,
  directed = FALSE,
  offset = TRUE
)
}
\arguments{
\item{n_nodes}{Number of nodes}

\item{K}{True number of groups}

\item{x_dim}{The number of covariates in X}

\item{prob}{Vector of probability of being in each group}

\item{beta0}{Intercept term, lower values indicate higher sparsity}

\item{beta}{Matrix of logistic regression coefficients}

\item{sigma}{Correlation matrix of the X matrix of covariates}

\item{spat}{Degree of spatial heterogeneity of z}

\item{directed}{Boolean indicating whether the network should be directed}

\item{offset}{Boolean indicating whether the degree terms theta should be used}
}
\value{
list of adjacency, membership, link probability, and distance
}
\description{
Generates network using probabilities, betas, and signal strength
}
