% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{gen_factor}
\alias{gen_factor}
\title{Helper function for CALF-SBM Initialization}
\usage{
gen_factor(initial_z, A, S_ij, directed = FALSE, offset = FALSE)
}
\arguments{
\item{initial_z}{Initial node membership vector}

\item{A}{Known adjacency matrix}

\item{S_ij}{Distance matrix}

\item{directed}{Boolean}

\item{offset}{Logistic regression offset terms (currently not used)}
}
\value{
List of X matrix and grid of (x, y) indices
}
\description{
Get the X matrix for logistic regression by taking clusters, links, and dist.
}
