% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{gen_factor}
\alias{gen_factor}
\title{Visualize Link Probability vs. Actual Matrix}
\usage{
gen_factor(initial_z, A, S_ij, directed = FALSE, offset = NULL)
}
\arguments{
\item{initial_z}{Initial node membership vector}

\item{A}{Known adjacency matrix}

\item{S_ij}{Distance matrix}

\item{directed}{logical; if \code{FALSE} (default), the MCMC output is from 
an undirected network}

\item{offset}{Logistic regression offset terms}
}
\value{
List of X matrix and grid of (x, y) indices
}
\description{
Helper function for CALF-SBM Initialization
}
\details{
Get the X matrix for logistic regression by taking clusters, links, and dist.
}
