% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{gen_factor}
\alias{gen_factor}
\title{Visualize Link Probability vs. Actual Matrix}
\usage{
gen_factor(initial_z, A, S_ij, directed = FALSE, offset = FALSE)
}
\arguments{
\item{initial_z}{Initial node membership vector}

\item{A}{Known adjacency matrix}

\item{S_ij}{Distance matrix}

\item{directed}{Boolean}

\item{offset}{Logistic regression offset terms (currently not used)}

\item{observed}{Observed adjacency matrix}

\item{prob}{Matrix of intrinsic probabilities}

\item{z_tru}{Vector of the true node membership}
}
\value{
Graph of the connectivity and graph of true probabilities

List of X matrix and grid of (x, y) indices
}
\description{
Using the adjacency matrix, returns a plot of the adjacency matrix, and 
a raster of the true probabilities
}
\note{
Internal helper function
}
\examples{
links <- gen_az(n_nodes = 50, K = 2, m = 2, prob = c(0.5, 0.5),
beta0 = 1, beta = diag(2) - 3, sigma = 0.3, spat = 0.5)
plot_conn(links$A, links$lp, links$z)
}
