% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{gen_x}
\alias{gen_x}
\title{(Obsolete) Helper Function to Generate Covariates}
\usage{
gen_x(x_dim, n_nodes, sigma)
}
\arguments{
\item{x_dim}{Dimension of X}

\item{n_nodes}{Number of nodes}

\item{sigma}{correlation (or correlation matrix if input is xdim by xdim)}
}
\value{
Matrix of covariates
}
\description{
Generate matrix of covariates X according to parameters
}
\note{
Deprecated
}
