% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{gibbs_sampler_unknown_k}
\alias{gibbs_sampler_unknown_k}
\title{(Obsolete) Unknown K Gibbs sampler}
\usage{
gibbs_sampler_unknown_k(dp, niter, A, S_ij, directed = FALSE)
}
\arguments{
\item{dp}{Positive number for CRP concentration parameter (recommended 1)}

\item{niter}{Total number of iterations to run}

\item{A}{Adjacency matrix}

\item{S_ij}{Distance matrix}

\item{directed}{logical; if \code{FALSE} (default), the output is from 
an undirected network}
}
\value{
List containing the estimated beta, node membership, and history
}
\description{
Gibbs sampler with CRP as initial state
}
\note{
Deprecated for now, but may be used in future
}
