% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{plot_conn}
\alias{plot_conn}
\title{Visualize Link Probability vs. Actual Matrix}
\usage{
plot_conn(observed, prob, z_tru)
}
\arguments{
\item{observed}{Observed adjacency matrix}

\item{prob}{Matrix of intrinsic probabilities}

\item{z_tru}{Vector of the true node membership}
}
\value{
Graph of the connectivity and graph of true probabilities
}
\description{
Using the adjacency matrix, returns a plot of the adjacency matrix, and 
a raster of the true probabilities
}
