% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{post_label_mcmc}
\alias{post_label_mcmc}
\title{Label-Switching According to Constraint}
\usage{
post_label_mcmc(mcmcSamples, K, directed = FALSE, lab = TRUE)
}
\arguments{
\item{mcmcSamples}{Matrix of raw MCMC samples, with each column corresponding 
to each beta value}

\item{K}{Number of groups}

\item{directed}{logical; if \code{FALSE} (default), the MCMC output is 
from an undirected network}

\item{lab}{logical; if \code{TRUE} (default), labels the clusters from 
smallest to largest mean values of within-cluster effects beta_{ii}}
}
\value{
Reorganized dataframe according to apparent beta_{ii} ordering
}
\description{
Function to post-label MCMC samples by ascending order of within-cluster 
values. This only orders it according the the means of the output. For 
label-switching at the sample level, see \code{post_label_mcmc_samples}
}
