% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{post_label_mcmc_samples}
\alias{post_label_mcmc_samples}
\title{Post-process raw MCMC output sample-by-sample}
\usage{
post_label_mcmc_samples(mcmcSamples, K, n, directed = FALSE)
}
\arguments{
\item{mcmcSamples}{Matrix of raw MCMC samples, with each column corresponding 
to each beta value}

\item{K}{Number of groups}

\item{n}{Total number of nodes.}

\item{directed}{logical; if \code{FALSE} (default), the MCMC output is from an 
undirected network}
}
\value{
Entire MCMC samples data reorganized according to constraint
}
\description{
Take a large matrix of raw MCMC samples from NIMBLE \code{MCMC}, and 
reorder each sample according to label-switching constraint
}
