% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{update_beta}
\alias{update_beta}
\title{Update Beta in Gibbs Sampler}
\usage{
update_beta(K, group, directed = FALSE, offset = FALSE)
}
\arguments{
\item{K}{Number of clusters}

\item{group}{Model matrix to be fitted on}

\item{directed}{Boolean indicating whether network is directed}

\item{offset}{Boolean indicating whether to use offset term}
}
\value{
beta0 and beta, with beta as a matrix
}
\description{
Helper function to update beta according to adjacency and node membership
}
