% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{update_beta_em}
\alias{update_beta_em}
\title{Update Beta in EM algorithm using logistic regression}
\usage{
update_beta_em(K, group, directed = FALSE, offset = FALSE, conf = 0.95)
}
\arguments{
\item{K}{A positive integer indicating the true number of clusters}

\item{group}{Model matrix to be fitted on}

\item{directed}{logical; if \code{FALSE} (default), the MCMC output is from 
an undirected network}

\item{offset}{Boolean indicating whether to use offset term}

\item{conf}{Confidence level of beta estimates (default = 0.95)}
}
\value{
beta0 and beta, with beta as a matrix
}
\description{
Helper function to update beta according to adjacency and node membership
}
\note{
Function \code{update_beta_bayes} is a helper in the 
in the \code{calfsbm_em} function, deriving an estimate for beta using 
the initial clustering configuration as input
}
