% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{update_z_from_beta}
\alias{update_z_from_beta}
\title{(Obsolete) Helper for Gibbs Sampler}
\usage{
update_z_from_beta(z, beta0, beta, S_ij, A, K, directed = FALSE)
}
\arguments{
\item{z}{Initial values of z}

\item{beta0}{Draw from posterior distribution}

\item{beta}{Draw from posterior distribution}

\item{S_ij}{Similarity or distance matrix}

\item{A}{Observed adjacency matrix}

\item{K}{Number of clusters}

\item{directed}{logical; if \code{FALSE} (default), the MCMC output is from 
an undirected network}
}
\value{
Updated node membership vector
}
\description{
Update node membership using likelihood
}
