% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_functions.R
\name{update_z_from_q}
\alias{update_z_from_q}
\title{Helper to update MFM-SBM by iteration}
\usage{
update_z_from_q(z, Q, A, conc)
}
\arguments{
\item{z}{Node membership}

\item{Q}{SBM probability matrix}

\item{A}{Known adjacency matrix}

\item{conc}{Concentration parameter}
}
\value{
Updated node membership
}
\description{
Helper function for MFM-SBM Algorithm 1
}
\note{
Internal helper
}
